수정사항 수정, test case 같은 경우는 아직 트랜잭션 생성 함수들이 미완성이라 작성후에 작성예정
