+++ service/blockchain/blockchain_test.go
@@ -16,17 +16,3 @@ func TestCreateNewBlockChainTest(t *testing.T){
 
 
 }
-
-func TestTxTest(t *testing.T) {
-	blk := Block{}
-	tx := Transaction{}
-	tx.txData = TxData{"ss","dd", Params{3, "func", []string{"10", "6"}}, "ff"}
-	tx.GenerateHash()
-
-	if blk.PutTranscation(tx) == true{
-		idx, err := blk.FindTransactionIndex(tx.transactionHash)
-		fmt.Println(idx, err)
-		fmt.Printf("%x\n", blk.transactions[idx].transactionHash)
-	}
-
-}
